implement new factory class
