+++ package.json
@@ -60,6 +60,7 @@
   },
   "dependencies": {
     "babel-core": "6.18.2",
+    "dot-prop": "^4.1.0",
     "eq3ble": "1.0.0",
     "is-mac": "1.0.1",
     "mqtt": "1.12.0",
