+++ src/callbackify.js
@@ -1,0 +1,9 @@
+export default function callbackify(fn) {
+  return (callback, ...args) => {
+    fn(...args).then((...results) => {
+      callback(null, ...results)
+    }, (err) => {
+      callback(err)
+    })
+  }
+}
