+++ src/constants.js
@@ -1,3 +1,5 @@
+import dotProp from 'dot-prop'
+
 export const TargetHeatingCoolingState = {
   OFF: 'TargetHeatingCoolingState.OFF',
   HEAT: 'TargetHeatingCoolingState.HEAT',
@@ -8,4 +10,8 @@ export const TargetHeatingCoolingState = {
 export const CurrentHeatingCoolingState = {
   OFF: 'CurrentHeatingCoolingState.OFF',
   HEAT: 'CurrentHeatingCoolingState.HEAT',
+}
+
+export function map(Characteristic, value) {
+  return dotProp.get(Characteristic, value)
 }
